requirements.txt
https://github.com/Rapptz/discord.py
discord.py
dnspython==1.16.0
async-timeout==3.0.1